Fix Accounts and Transactions view recursion binding error
When an account or transaction was selected within the datagrid and an edit or delete button was pressed, the edit and delete buttons would become disabled, despite still having a selected account or transaction. The binding mode was set to two way for the edit and delete dialog buttons, which didn't need to be. If you tried to set the HasSelectedItem to any value within the viewmodel, a StackOverFlow exception would appear. Changing the binding mode to OneWay solved the issue.